Time for a break.
I've managed to destroy two virtualenv's, find (and lose) a
Mystery Image that makes pyglet crash on load, and not submit
a bug report about it.

On the up side, the basic framework _should_ be in place
to use pyglet, and it's pretty simple.  I should just need
to shim out the rest of the pygame-dependent code, and then
we can start replacing it bit by bit.