Things draw now!
Even if they draw wrong.  There's some silliness with
coordinate systems happening.

Also we can't draw primitive circles and lines, so we
need some sort of workaround for that.  That was a bit
of a hack anyway; circles are placeholdered in, just
need to figure out how to do lines.  Should be possible.

Next up, I think, is input.