Not much done, starting to think about gamestate organization.
Basically, we need to pass the gamestate to each object when
it recieves an event.  The Pyglet event handler loop can't do
this.  Do we:

a) Write our own event handler loop
b) Make each object keep a reference to the gamestate anyway
or c) make the gamestate a global?

I sort of lean towards (a).  Dragon sort of leans toward (b).
(c) is certainly valid.  Sigh.